feat: provide an extendable template registry
