+++ addon/components/breadcrumbs-container.hbs
@@ -1,0 +1,9 @@
+<ul
+  {{this.breadcrumbsContainerModifier
+    itemClass=@itemClass
+    linkClass=@linkClass
+  }}
+  ...attributes
+>
+  {{yield}}
+</ul>
