--- addon/components/breadcrumbs-container/index.hbs
@@ -1,6 +1,0 @@
-<ul
-  {{breadcrumbs-container itemClass=@itemClass linkClass=@linkClass}}
-  ...attributes
->
-  {{yield}}
-</ul>
