--- addon/components/breadcrumbs-container/index.ts
@@ -1,14 +1,0 @@
-import templateOnlyComponent from '@ember/component/template-only';
-
-interface BreadcrumbsContainerComponentSignature {
-  Args: {
-    itemClass?: string;
-    linkClass?: string;
-  };
-  Blocks: {
-    default: [];
-  };
-  Element: HTMLUListElement;
-}
-
-export default templateOnlyComponent<BreadcrumbsContainerComponentSignature>();
