--- app/modifiers/breadcrumbs-container.js
@@ -1,1 +1,0 @@
-export { default } from '@bagaar/ember-breadcrumbs/modifiers/breadcrumbs-container';
